put tab-over mode out of its misery
this commit started its life with much more green.

having spent an entire day steadily implementing the blasted thing, it turns out
that the sort of access content scripts have to page scopes is just not enough.
perhaps some day in the future we can reconsider this after writing a TW plugin
that can pick up on postMessage calls or something, but as it stands, there just
isn't enough power in executeScript(). webserver mode is so good that the only
part of this that honestly hurts is the wasted time.

don't write browser code, people