storage.js: add created/modified in default format
for some reason this stopped working in webserver mode with a recent tw update.
this should have been the default to accommodate for download mode regardless;
dateless tiddlers are no fun.

in future releases it would be good to provide `now` as a variable to avoid
spamming the user with freaky js in the settings page